+++ src/dif/clogic/app/MelodyListActivity.java
@@ -1,31 +1,33 @@
 package dif.clogic.app;
 
+import android.app.ActionBar;
 import android.app.Activity;
 import android.app.AlertDialog;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.Cursor;
+import android.graphics.Typeface;
 import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
 import android.view.LayoutInflater;
-import android.view.Menu;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ListView;
+import android.widget.*;
+import com.leff.midi.MidiFile;
+import com.leff.midi.MidiTrack;
+import com.leff.midi.event.ChannelEvent;
+import com.leff.midi.event.MidiEvent;
+import com.leff.midi.event.ProgramChange;
 import dif.clogic.custom.CustomProgressBar;
-import dif.clogic.other.Accompaniment;
-import dif.clogic.other.DbOpenHelper;
-import dif.clogic.other.Melody;
-import dif.clogic.other.MelodyAdapter;
+import dif.clogic.other.*;
 
 import java.io.File;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Iterator;
 
 /**
  * Created with IntelliJ IDEA.
@@ -40,11 +42,9 @@ public class MelodyListActivity extends Activity {
     private MelodyAdapter melodyAdapter;
     private MediaPlayer mPlayer = null;
     private DbOpenHelper mDbOpenHelper;
-    private Button addButton;
     private ListView listView;
 
-    //private SeekBar playerSeekBar;
-    private Button playButton;
+    private ImageButton playButton;
 
     private CustomProgressBar progress;
 
@@ -54,7 +54,7 @@ public class MelodyListActivity extends Activity {
 
         setTitle("멜로디 리스트");
 
-        /*playButton = (Button)findViewById(R.id.melodyPlayButton);
+        playButton = (ImageButton)findViewById(R.id.melodyPlayButton);
         playButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -63,41 +63,22 @@ public class MelodyListActivity extends Activity {
                     return;
 
                 if(mPlayer.isPlaying()) {
-                    playButton.setText("재생");
+                    playButton.setImageDrawable(getResources().getDrawable(R.drawable.play5));
                     mPlayer.pause();
                 } else {
-                    playButton.setText("일시정지");
+                    playButton.setImageDrawable(getResources().getDrawable(R.drawable.pause2));
                     mPlayer.start();
                 }
             }
-        });*/
+        });
 
         mPlayer = new MediaPlayer();
 
+        TextView t = (TextView)findViewById(R.id.playingTitle);
+        Typeface font = Typeface.createFromAsset(getAssets(), "nanumN.ttf");
+        t.setTypeface(font);
+
         progress = (CustomProgressBar)findViewById(R.id.melodyProgress);
-
-        /*playerSeekBar = (SeekBar)findViewById(R.id.melodySeekBar);
-        playerSeekBar.setVisibility(ProgressBar.VISIBLE);
-        playerSeekBar.setProgressDrawable(this.getResources().getDrawable(R.drawable.custom_seekbar));
-        playerSeekBar.setThumb(this.getResources().getDrawable(R.drawable.thumb));
-        playerSeekBar.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-            @Override
-            public void onProgressChanged(SeekBar seekBar, int i, boolean b) {
-                //To change body of implemented methods use File | Settings | File Templates.
-            }
-
-            @Override
-            public void onStartTrackingTouch(SeekBar seekBar) {
-                //To change body of implemented methods use File | Settings | File Templates.
-            }
-
-            @Override
-            public void onStopTrackingTouch(SeekBar seekBar) {
-                //To change body of implemented methods use File | Settings | File Templates.
-                if (mPlayer != null)
-                    mPlayer.seekTo(seekBar.getProgress());
-            }
-        });*/
 
         new Thread(new Runnable() {
             @Override
@@ -109,7 +90,6 @@ public class MelodyListActivity extends Activity {
 
                     if(mPlayer.isPlaying())
                         progress.setProgress(mPlayer.getCurrentPosition());
-                        //playerSeekBar.setProgress(mPlayer.getCurrentPosition());
                 }
             }
         }).start();
@@ -130,21 +110,8 @@ public class MelodyListActivity extends Activity {
         } else {
         }
 
-        /*addButton = (Button)findViewById(R.id.addAccompanimentBtn);
-        addButton.setText("멜로디 그리기");
-        addButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                //To change body of implemented methods use File | Settings | File Templates.
-                // go to drawActivity!!
-                Intent intent = new Intent(MelodyListActivity.this, CheckboxListViewActivity.class);
-                startActivity(intent);
-            }
-        });*/
-
-        listView = (ListView)findViewById(R.id.accompanimentListView);
+        listView = (ListView)findViewById(R.id.melodyListView);
         listView.setAdapter(melodyAdapter);
-        listView.setDividerHeight(0);
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, final int which, long l) {
@@ -164,19 +131,15 @@ public class MelodyListActivity extends Activity {
                             if (mPlayer == null)
                                 mPlayer = new MediaPlayer();
 
-                            if(mPlayer.isPlaying()) {
-                                mPlayer.stop();
-                                mPlayer.reset();
-                            } else {
-                                mPlayer.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
-                                    @Override
-                                    public void onCompletion(MediaPlayer mediaPlayer) {
-                                        //To change body of implemented methods use File | Settings | File Templates.
-                                        //playerSeekBar.setProgress(0);
-                                        progress.setProgress(0);
-                                    }
-                                });
-                            }
+                            mPlayer.reset();
+                            mPlayer.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
+                                @Override
+                                public void onCompletion(MediaPlayer mediaPlayer) {
+                                    //To change body of implemented methods use File | Settings | File Templates.
+                                    progress.setProgress(0);
+                                    playButton.setImageDrawable(getResources().getDrawable(R.drawable.play5));
+                                }
+                            });
 
                             try {
                                 String str = Environment.getExternalStorageDirectory().getAbsolutePath();
@@ -190,17 +153,66 @@ public class MelodyListActivity extends Activity {
                             }
                             mPlayer.start();
 
-                            playButton.setText("일시정지");
+                            playButton.setImageDrawable(getResources().getDrawable(R.drawable.pause2));
 
-                            /*playerSeekBar.setProgress(0);
-                            playerSeekBar.setMax(mPlayer.getDuration());*/
                             progress.setProgress(0);
                             progress.setMax(mPlayer.getDuration());
+
+                            TextView t = (TextView)findViewById(R.id.playingTitle);
+                            t.setText(melodyList.get(which).Name);
+                            Typeface font = Typeface.createFromAsset(getAssets(), "nanumN.ttf");
+                            t.setTypeface(font);
+                        }
+
+                        if (data[i].equals("악기 설정")) {
+                            AlertDialog.Builder builder = new AlertDialog.Builder(MelodyListActivity.this);
+                            builder.setTitle("악기 설정");
+                            builder.setItems(Instruments.instruments, new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialogInterface, int i) {
+                                    //To change body of implemented methods use File | Settings | File Templates.
+                                    File file = new File(Environment.getExternalStorageDirectory().getAbsolutePath() + File.separator + melodyList.get(which).Name + ".mid");
+
+                                    MidiFile mFile = null;
+                                    try {
+                                        mFile = new MidiFile(file);
+                                    } catch (IOException e) {
+                                        e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
+                                    }
+                                    if(mFile == null)
+                                        return;
+
+                                    MidiTrack t = mFile.getTracks().get(1);
+
+                                    String[] instruments = getResources().getStringArray(R.array.instruments);
+
+                                    Iterator<MidiEvent> it = t.getEvents().iterator();
+                                    ArrayList<MidiEvent> removeProgramChange = new ArrayList<MidiEvent>();
+                                    while(it.hasNext()) {
+                                        ChannelEvent event = (ChannelEvent) it.next();
+                                        if(event.getType() == ChannelEvent.PROGRAM_CHANGE)
+                                            removeProgramChange.add(event);
+                                    }
+
+                                    for(MidiEvent e : removeProgramChange) {
+                                        t.removeEvent(e);
+                                    }
+
+                                    ProgramChange pc = new ProgramChange(0, 0, 0);
+                                    pc.setProgramNumber(i);
+                                    t.insertEvent(pc);
+
+                                    try {
+                                        mFile.writeToFile(file);
+                                    } catch (IOException e) {
+                                        e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
+                                    }
+                                }
+                            });
+                            builder.show();
                         }
 
                         if (data[i].equals("악보 보기")) {
-
-                            //Cursor cursor = mDbOpenHelper.getMatchName("example");
                             String params = Accompaniment.convert(melodyList.get(which).originRecord);
                             Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse("http://druzicofclogic.appspot.com/melody/view/" + params)); // CN5 같은 쿼리 보내주기~~
                             startActivity(intent);
@@ -223,7 +235,8 @@ public class MelodyListActivity extends Activity {
 
                                             File file = new File(Environment.getExternalStorageDirectory().getAbsolutePath() + File.separator + beforeText + ".mid");
                                             File to = new File(Environment.getExternalStorageDirectory().getAbsolutePath() + File.separator + afterText + ".mid");
-                                            file.renameTo(to);
+                                            if(!to.exists())
+                                                file.renameTo(to);
 
                                             melodyList.get(which).Name = afterText;
 
@@ -262,12 +275,28 @@ public class MelodyListActivity extends Activity {
                 builder.show();
             }
         });
-    }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.addbtn, menu);
-        return true;
+        ActionBar ab = getActionBar();
+        ab.setDisplayShowCustomEnabled(true);
+        ab.setDisplayShowTitleEnabled(false);
+        LayoutInflater inflater = (LayoutInflater)this.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        View v = inflater.inflate(R.layout.custom_actionbar, null);
+        TextView tv = (TextView)v.findViewById(R.id.actionBarTitle);
+        tv.setTypeface(font);
+
+        ImageButton button = (ImageButton)v.findViewById(R.id.addBtn);
+        button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                //To change body of implemented methods use File | Settings | File Templates.
+                Intent intent = new Intent(MelodyListActivity.this, CheckboxListViewActivity.class);
+                startActivity(intent);
+            }
+        });
+
+        ab.setCustomView(v);
+        ab.setDisplayShowHomeEnabled(false);
+        ab.setDisplayShowTitleEnabled(false);
     }
 
     @Override
@@ -307,6 +336,7 @@ public class MelodyListActivity extends Activity {
                     cursor.getString(cursor.getColumnIndex("name")),
                     cursor.getString(cursor.getColumnIndex("originrecord")),
                     cursor.getString(cursor.getColumnIndex("accompanimentlist"))));
+
         }
     }
 }
